requirements.txt
keras==2.3.1
tensorflow==2.0.1
scikit-learn
numpy~=1.18.5
matplotlib~=3.3.2
opencv-python~=4.4.0.46
imutils~=0.5.3
scikit-image~=0.17.2
Pillow~=8.0.1
PyQt5~=5.15.2
pytesseract~=0.3.6
loremipsum~=1.0.5